Fix ESLint errors
