+++ plugins/renamer/src/search/result_processors/clean_up_result.ts
@@ -6,7 +6,7 @@ const WORDS_IGNORING_CAPITALIZATION = ["at"];
 
 export function cleanUpResult(result: Result): string {
   return result.title
-    .replace(/\_|\-|\#/g, " ")
+    .replace(/_|-|#/g, " ")
     .split(" ")
     .map((word) => {
       if (WORDS_IGNORING_CAPITALIZATION.includes(word)) return word;
